feat: Enable the native installation of tensorflow on arm64.
Due to the installation problem of TensorFlow in Docker images for
arm64 platform, this workspace was forced to build and execute under
linux/amd64 mode. However, this workaround further causes a special
error during running Chisel toolchains with sbt, which is probably led
by some problems of Rosetta.

To deal with this problem, this commit successfully installs TensorFlow
in arm64 images. With this update, the workspace no longer needs to be
built as a linux/amd64 image on arm64 machines, and the Chisel error
can also be solved directly.

Besides, with the version of the base image is updated to ubuntu:22.04,
the default version of Python is also updated to 3.10. As a result, the
conda package is removed to make the workspace more lightweight since
the Python 3.9 environment is not required anymore.